Updated Pick to PickModel
